+++ randovania/games/common/prime_family/exporter/good_hashes.py
@@ -1,0 +1,12 @@
+PRIME1_GC_ISOS = (
+    "sha1:ac20c744db18fdf0339f37945e880708fd317231",  # NTSC 0-00
+    "sha1:4ba8933499e0b74b2f6006d622e4fbc7593ab3c7",  # NTSC 0-01
+    "sha1:1a737910b55b59c6ad91be9e3e3c43517fd52efb",  # NTSC 0-02
+    "sha1:34ac8a764a3c1db3326c39071cee2fc49e730aca",  # PAL
+    "sha1:ee6c58b46012ebffb615506ec43e02ba71905662",  # Japan
+)
+
+PRIME2_GC_ISOS = (
+    "sha1:6af3496391bfe4ae97769f2fe85f0f575f2e3d5a",  # NTSC
+    "sha1:a779afd7c705743e003a21412ad6bc8be345fbad",  # PAL
+)