+++ randovania/games/prime2/exporter/game_exporter.py
@@ -12,6 +12,7 @@ from retro_data_structures.game_check import Game as RDSGame
 
 from randovania import monitoring
 from randovania.exporter.game_exporter import GameExporter, GameExportParams
+from randovania.games.common.prime_family.exporter import good_hashes
 from randovania.games.prime2.exporter.claris_randomizer_data import decode_randomizer_data
 from randovania.games.prime2.exporter.export_params import EchoesGameExportParams
 from randovania.games.prime2.exporter.patch_data_factory import adjust_model_name
@@ -53,6 +54,12 @@ class EchoesGameExporter(GameExporter):
 
     def _after_export(self):
         self._busy = False
+
+    def known_good_hashes(self) -> dict[str, tuple[str, ...]]:
+        return {
+            "prime1_iso": good_hashes.PRIME1_GC_ISOS,
+            "prime2_iso": good_hashes.PRIME2_GC_ISOS,
+        }
 
     def _do_export_game(
         self,